prototype/requirements.txt
torch
requests
pillow
diffusers
gradio
numpy
tqdm